Remove Kubernetes Python example
This example uses an earlier version of the Kubernetes provider, which
was built atop terraform. We've moved to a new implementation of
Kubernetes support and we haven't built out the python support yet.

Delete this out of date example, we can bring it back once we have
Kubernetes working again with Python.