Use explicit namespace in sock shop
What we have now is a direct port of the canonical Sock Shop Kubernetes
example. This example does not have the sock-shop namespace, but does
use it, causing the example to crash out of the box.

Here we fix that by simply declaring and using the namespace explicitly.

We also have all services explicitly reference the deployments they
depend on.