Use "latest" as the version in more examples
For a few of the examples we had harnessed, they were using older
versions of our packages. Move them to "latest" so our examples always
reflect the most recently released versions of our packages.

There are no lock file updates as part of this change as we do not
check yarn.lock's into git here.