Upgrade `pip` inside container
We are seeing some issues in CI where `awscli` fails to install with
errors deep in the bowels of pip. I'm unable to reproduce this
locally, but I suspect that there's an network issue in Travis that is
causing us to need to retry downloads, which is hitting some bad path
inside the very old version of `pip` we have

Let's try upgrading it to the latest version (19.X vs 8.X!) and see if
it helps.